+++ system/defines.php
@@ -9,7 +9,7 @@
 
 // Some standard defines
 define('GRAV', true);
-define('GRAV_VERSION', '1.7.53.1');
+define('GRAV_VERSION', '1.7.53.2');
 define('GRAV_SCHEMA', '1.7.53_2026-06-09_0');
 define('GRAV_TESTING', false);